+++ database/migrations/2025_02_20_111606_add_role_to_users_table.php
@@ -12,10 +12,16 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->enum('role_type', ['admin', 'user_metier', 'fournisseur_morale', 'fournisseur_physique'])
-            ->nullable()
-            ->after('id');
-            $table->unsignedBigInteger('role_id')->nullable()->after('role_type');
+            if (!Schema::hasColumn('users', 'role_type')) {
+                $table->enum('role_type', ['admin', 'user_metier', 'fournisseur_morale', 'fournisseur_physique'])
+                      ->nullable()
+                      ->after('id');
+            }
+
+            if (!Schema::hasColumn('users', 'role_id')) {
+                $table->unsignedBigInteger('role_id')->nullable()->after('role_type');
+            }
+
             $table->index(['role_type', 'role_id']);
         });
     }
@@ -27,8 +33,14 @@ return new class extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             $table->dropIndex(['role_type', 'role_id']);
-            $table->dropColumn('role_type');
-            $table->dropColumn('role_id');
+
+            if (Schema::hasColumn('users', 'role_type')) {
+                $table->dropColumn('role_type');
+            }
+
+            if (Schema::hasColumn('users', 'role_id')) {
+                $table->dropColumn('role_id');
+            }
         });
     }
 };
